Refactor tracking functions to static methods
