+++ tbl_tracking.php
@@ -5,14 +5,12 @@
  *
  * @package PhpMyAdmin
  */
-
-// Run common work
+use PhpMyAdmin\Message;
 use PhpMyAdmin\Tracker;
+use PhpMyAdmin\Tracking;
 use PhpMyAdmin\Response;
 
 require_once './libraries/common.inc.php';
-
-require_once './libraries/tracking.lib.php';
 
 //Get some js files needed for Ajax requests
 $response = Response::getInstance();
@@ -31,7 +29,7 @@ if (Tracker::isActive()
     && ! (isset($_REQUEST['report_export'])
     && $_REQUEST['export_type'] == 'sqldumpfile')
 ) {
-    $msg = PhpMyAdmin\Message::notice(
+    $msg = Message::notice(
         sprintf(
             __('Tracking of %s is activated.'),
             htmlspecialchars($GLOBALS["db"] . '.' . $GLOBALS["table"])
@@ -80,14 +78,14 @@ if (isset($_REQUEST['report']) || isset($_REQUEST['report_export'])) {
 
 // Prepare export
 if (isset($_REQUEST['report_export'])) {
-    $entries = PMA_getEntries($data, $filter_ts_from, $filter_ts_to, $filter_users);
+    $entries = Tracking::getEntries($data, $filter_ts_from, $filter_ts_to, $filter_users);
 }
 
 // Export as file download
 if (isset($_REQUEST['report_export'])
     && $_REQUEST['export_type'] == 'sqldumpfile'
 ) {
-    PMA_exportAsFileDownload($entries);
+    Tracking::exportAsFileDownload($entries);
 }
 
 $html = '<br />';
@@ -99,59 +97,59 @@ if (isset($_REQUEST['submit_mult'])) {
     if (! empty($_REQUEST['selected_versions'])) {
         if ($_REQUEST['submit_mult'] == 'delete_version') {
             foreach ($_REQUEST['selected_versions'] as $version) {
-                PMA_deleteTrackingVersion($version);
+                Tracking::deleteTrackingVersion($version);
             }
-            $html .= PhpMyAdmin\Message::success(
+            $html .= Message::success(
                 __('Tracking versions deleted successfully.')
             )->getDisplay();
         }
     } else {
-        $html .= PhpMyAdmin\Message::notice(
+        $html .= Message::notice(
             __('No versions selected.')
         )->getDisplay();
     }
 }
 
 if (isset($_REQUEST['submit_delete_version'])) {
-    $html .= PMA_deleteTrackingVersion($_REQUEST['version']);
+    $html .= Tracking::deleteTrackingVersion($_REQUEST['version']);
 }
 
 // Create tracking version
 if (isset($_REQUEST['submit_create_version'])) {
-    $html .= PMA_createTrackingVersion();
+    $html .= Tracking::createTrackingVersion();
 }
 
 // Deactivate tracking
 if (isset($_REQUEST['toggle_activation'])
     && $_REQUEST['toggle_activation'] == 'deactivate_now'
 ) {
-    $html .= PMA_changeTracking('deactivate');
+    $html .= Tracking::changeTracking('deactivate');
 }
 
 // Activate tracking
 if (isset($_REQUEST['toggle_activation'])
     && $_REQUEST['toggle_activation'] == 'activate_now'
 ) {
-    $html .= PMA_changeTracking('activate');
+    $html .= Tracking::changeTracking('activate');
 }
 
 // Export as SQL execution
 if (isset($_REQUEST['report_export']) && $_REQUEST['export_type'] == 'execution') {
-    $sql_result = PMA_exportAsSQLExecution($entries);
-    $msg = PhpMyAdmin\Message::success(__('SQL statements executed.'));
+    $sql_result = Tracking::exportAsSqlExecution($entries);
+    $msg = Message::success(__('SQL statements executed.'));
     $html .= $msg->getDisplay();
 }
 
 // Export as SQL dump
 if (isset($_REQUEST['report_export']) && $_REQUEST['export_type'] == 'sqldump') {
-    $html .= PMA_exportAsSQLDump($entries);
+    $html .= Tracking::exportAsSqlDump($entries);
 }
 
 /*
  * Schema snapshot
  */
 if (isset($_REQUEST['snapshot'])) {
-    $html .= PMA_getHtmlForSchemaSnapshot($url_query);
+    $html .= Tracking::getHtmlForSchemaSnapshot($url_query);
 }
 // end of snapshot report
 
@@ -161,11 +159,11 @@ if (isset($_REQUEST['snapshot'])) {
 if (isset($_REQUEST['report'])
     && (isset($_REQUEST['delete_ddlog']) || isset($_REQUEST['delete_dmlog']))
 ) {
-    $html .= PMA_deleteTrackingReportRows($data);
+    $html .= Tracking::deleteTrackingReportRows($data);
 }
 
 if (isset($_REQUEST['report']) || isset($_REQUEST['report_export'])) {
-    $html .= PMA_getHtmlForTrackingReport(
+    $html .= Tracking::getHtmlForTrackingReport(
         $url_query, $data, $url_params, $selection_schema, $selection_data,
         $selection_both, $filter_ts_to, $filter_ts_from, $filter_users
     );
@@ -175,9 +173,9 @@ if (isset($_REQUEST['report']) || isset($_REQUEST['report_export'])) {
 /*
  * List selectable tables
  */
-$selectable_tables_sql_result = PMA_getSQLResultForSelectableTables();
+$selectable_tables_sql_result = Tracking::getSqlResultForSelectableTables();
 if ($GLOBALS['dbi']->numRows($selectable_tables_sql_result) > 0) {
-    $html .= PMA_getHtmlForSelectableTables(
+    $html .= Tracking::getHtmlForSelectableTables(
         $selectable_tables_sql_result, $url_query
     );
 }
@@ -186,10 +184,10 @@ $html .= '<br />';
 /*
  * List versions of current table
  */
-$sql_result = PMA_getListOfVersionsOfTable();
-$last_version = PMA_getTableLastVersionNumber($sql_result);
+$sql_result = Tracking::getListOfVersionsOfTable();
+$last_version = Tracking::getTableLastVersionNumber($sql_result);
 if ($last_version > 0) {
-    $html .= PMA_getHtmlForTableVersionDetails(
+    $html .= Tracking::getHtmlForTableVersionDetails(
         $sql_result, $last_version, $url_params,
         $url_query, $pmaThemeImage, $text_dir
     );
@@ -197,7 +195,7 @@ if ($last_version > 0) {
 
 $type = $GLOBALS['dbi']->getTable($GLOBALS['db'], $GLOBALS['table'])
     ->isView() ? 'view' : 'table';
-$html .= PMA_getHtmlForDataDefinitionAndManipulationStatements(
+$html .= Tracking::getHtmlForDataDefinitionAndManipulationStatements(
     'tbl_tracking.php' . $url_query,
     $last_version,
     $GLOBALS['db'],
